third_party/blink/web_tests/fullscreen/full-screen-request-not-supported.html

<body>
<script src="full-screen-test.js"></script>
<video id="video" width="300"></video>
<div>Tests for the fullscreenSupported setting which will be set to false
in those platforms where fullscreen is not supported.
<script>
    if (window.internals) {
        internals.settings.setFullscreenSupported(false);

        waitForEvent(document, 'webkitfullscreenchange', function() {
            logResult(false, 'Entered fullscreen.');
            endTest();
        });
        waitForEventOnceAndEnd(document, 'webkitfullscreenerror');
        runWithKeyDown(function(){document.getElementById('video').webkitRequestFullScreen()});
    }
</script>
</body>
