third_party/blink/web_tests/fullscreen/full-screen-request-not-supported-expected.txt

EVENT(webkitfullscreenerror)
END OF TEST

Tests for the fullscreenSupported setting which will be set to false in those platforms where fullscreen is not supported.
